+++ packages/apollo/src/commands/client/codegen.ts
@@ -41,7 +41,8 @@ export default class Generate extends ClientCommand {
     }),
     localSchemaFile: flags.string({
       description:
-        "Path to your local GraphQL schema file (introspection result or SDL)"
+        "Path to your local GraphQL schema file (introspection result or SDL)",
+      multiple: true
     }),
     addTypename: flags.boolean({
       description: