+++ packages/apollo/src/commands/service/push.ts
@@ -19,7 +19,8 @@ export default class ServicePush extends ProjectCommand {
     }),
     localSchemaFile: flags.string({
       description:
-        "Path to your local GraphQL schema file (introspection result or SDL)"
+        "Path to your local GraphQL schema file (introspection result or SDL)",
+      multiple: true
     }),
     federated: flags.boolean({
       char: "f",